+++ MiDAS/R/MiDAS_stats.R
@@ -329,6 +329,10 @@ prepareHlaData <- function(hla_calls,
 #' @inheritParams analyzeAssociations
 #' @inheritParams analyzeConditionalAssociations
 #' @inheritParams formatAssociationsResults
+#' @param analysis_type String indicating the type of analysis being performed,
+#'   at this point it is only used for results formatting. Valid values are
+#'   \code{"hla_alleles"}, \code{"aa_level"}, \code{"expression_levels"},
+#'   \code{"allele_groups"}, \code{"custom"}.
 #' @param conditional Logical indicating if the analysis should be performed
 #'   using stepwise conditional tests or not. See
 #'   \link{analyzeConditionalAssociations} for more details.
@@ -340,12 +344,20 @@ prepareHlaData <- function(hla_calls,
 #'   this number will not be considered during analysis. If it's greater or
 #'   equal 1 variables with number of counts less that this will not be
 #'   considered during analysis.
+#' @param logistic Logical indicating if statistical model used is logistic (eg.
+#'   \code{coxph}). If \code{NULL} function will try to figure this out. This is
+#'   only used for results formatting.
+#' @param binary_phenotype Logical indicating if coefficient estimates should be
+#'   exponentiated. This is typical for logistic and multinomial regressions,
+#'   but a bad idea if there is no log or logit link. If \code{NULL} function
+#'   will try to figure this out by testing if response is binary (\code{0} or
+#'   \code{1}).
 #' @param kable_output Logical indicating if additionally results should be
 #'   pretty printed in specified \code{format}.
 #'
-#' \code{variables} takes \code{NULL} as a default value. When specified as such
-#' column names of data frame associated with the \code{object} are used as
-#' variables for testing. This excludes first column which should corresponds
+#' \code{variables} takes \code{NULL} as a default value, than column names of
+#' data frame associated with the \code{object} are used as variables in
+#' association tests. This excludes first column which should corresponds
 #' to samples IDs as well as covariates and response variables defined in
 #' object formula.
 #'
@@ -363,14 +375,15 @@ prepareHlaData <- function(hla_calls,
 #' pheno <- read.table(pheno_file, header = TRUE, stringsAsFactors = FALSE)
 #' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
 #' covar <- read.table(covar_file, header = TRUE, stringsAsFactors = FALSE)
-#' midas_data <- prepareHlaData(hla_calls = hla_calls,
-#'                              pheno = pheno,
-#'                              covar = covar,
-#'                              inheritance_model = "additive"
+#' midas_data <- prepareMiDASData(hla_calls = hla_calls,
+#'                                pheno = pheno,
+#'                                covar = covar,
+#'                                analysis_type = "hla_allele",
+#'                                inheritance_model = "additive"
 #' )
 #'
 #' object <- coxph(Surv(OS, OS_DIED) ~ AGE + SEX, data = midas_data)
-#' analyzeMiDASData(object)
+#' analyzeMiDASData(object, analysis_type = "hla_alleles")
 #'
 #' @importFrom assertthat assert_that is.flag is.number is.string
 #' @importFrom dplyr filter left_join select rename
@@ -380,15 +393,17 @@ prepareHlaData <- function(hla_calls,
 #'
 #' @export
 analyzeMiDASData <- function(object,
+                             analysis_type = c("hla_alleles", "aa_level", "expression_levels",  "allele_groups", "custom"),
                              conditional = FALSE,
                              variables = NULL,
                              frequency_cutoff = NULL,
                              pvalue_cutoff = NULL,
                              correction = "BH",
+                             logistic = NULL,
+                             binary_phenotype = NULL,
                              th = 0.05,
                              rss_th = 1e-07,
                              kable_output = TRUE,
-                             type = "hla_alleles",
                              format = getOption("knitr.table.format")) {
 
   assert_that(
@@ -400,6 +415,10 @@ analyzeMiDASData <- function(object,
   object_variables <- colnames(object_data)[-1]
 
   assert_that(
+    is.string(analysis_type),
+    stringMatches(analysis_type,
+                  choice = c("hla_alleles", "aa_level", "expression_levels",  "allele_groups", "custom")
+    ),
     is.flag(conditional),
     isCharacterOrNULL(variables),
     see_if(
@@ -411,13 +430,11 @@ analyzeMiDASData <- function(object,
     isNumberOrNULL(frequency_cutoff),
     isNumberOrNULL(pvalue_cutoff),
     is.string(correction),
+    isFlagOrNULL(logistic),
+    isFlagOrNULL(binary_phenotype),
     is.number(th),
     is.number(rss_th),
     is.flag(kable_output),
-    is.string(type),
-    stringMatches(type,
-                  choice = c("hla_alleles", "aa_level", "expression_levels")
-    ),
     is.string(format),
     stringMatches(format, choice = c("html", "latex"))
   )
@@ -430,11 +447,13 @@ analyzeMiDASData <- function(object,
     )
   }
 
-  # guess if model used is logistic type -- so far the logistic models in HLA analysis I've seen are coxph and glm(family = binomial)
-  model_fun <- deparse(object_call[[1]])
-  model_family <- deparse(object_call[["family"]])
-  logistic <- grepl("coxph", model_fun) |
-    (grepl("glm", model_fun) & grepl("binomial", model_family))
+  # guess if model used is logistic type
+  if (is.null(logistic)) {
+    model_fun <- deparse(object_call[[1]])
+    model_family <- deparse(object_call[["family"]])
+    logistic <- grepl("coxph", model_fun) |
+       (grepl("glm", model_fun) & grepl("binomial", model_family))
+  }
 
   # Filter variables on frequency cutoff
   frequency_cutoff <- ifelse(is.null(frequency_cutoff), 0, frequency_cutoff)
@@ -467,8 +486,12 @@ analyzeMiDASData <- function(object,
   results <- left_join(x = results, y = variables_freq, by = "term")
 
   pheno_var <- all.vars(object_formula)[1]
-  binary_phenotype <- object_data[, pheno_var] %in% c(0, 1)
-  if (all(binary_phenotype, na.rm = TRUE)) {
+  if (is.null(binary_phenotype)) {
+    binary_phenotype <- object_data[, pheno_var] %in% c(0, 1)
+    binary_phenotype <- all(binary_phenotype, na.rm = TRUE)
+  }
+
+  if (binary_phenotype) {
     results <- object_data %>%
       filter(.data[[!! pheno_var]] == 1) %>%
       select("ID", !! variables) %>%
@@ -491,7 +514,7 @@ analyzeMiDASData <- function(object,
 
     preety_table <- formatAssociationsResults(
       results,
-      type = type,
+      type = analysis_type,
       response_variable = response_variable,
       logistic = logistic,
       pvalue_cutoff = pvalue_cutoff,
@@ -502,10 +525,11 @@ analyzeMiDASData <- function(object,
 
   # rename term and estimate to match preety_table
   estimate_name <- ifelse(logistic, "odds.ratio", "estimate")
-  term_name <- switch (type,
+  term_name <- switch (analysis_type,
                        "hla_alleles" = "allele",
                        "aa_level" = "aa",
                        "expression_levels" = "allele",
+                       "allele_groups" = "allele.group",
                        "term"
   )
 
@@ -513,4 +537,211 @@ analyzeMiDASData <- function(object,
     rename(!! term_name := .data$term, !! estimate_name := .data$estimate)
 
   return(results)
+}
+
+#' Prepare HLA calls data for statistical analysis
+#'
+#' \code{prepareMiDASData} transform HLA alleles calls according to selected
+#' analysis type and joins obtained transformation with additional data frames
+#' like phenotypic observations or covariates, creating an input data for
+#' further statistical analysis.
+#'
+#' @inheritParams checkHlaCallsFormat
+#' @inheritParams hlaCallsToCounts
+#' @inheritParams hlaToAAVariation
+#' @param ... Data frames holding additional variables like phenotypic
+#'   observations or covariates.
+#' @param analysis_type String indicating analysis type for which data should be
+#'   prepared. Valid choices are \code{"hla_allele"}, \code{"aa_level"},
+#'   \code{"expression_levels"}, \code{"allele_groups"}, \code{"custom"}. See
+#'   details for further explanations.
+#'
+#' \code{...} should be data frames with first column holding samples IDs and
+#' named \code{ID}. Those should correspond to \code{ID} column in
+#' \code{hla_calls}.
+#'
+#' \code{analysis_type} specifies type of analysis for which \code{hla_calls}
+#' should be prepared:
+#'
+#' \code{"hla_allele"} - \code{hla_calls} are transformed into counts  under
+#' \code{inheritance_model} of choice (this is done with
+#' \link{hlaCallsToCounts}).
+#'
+#' \code{"aa_level"} - \code{hla_calls} are first converted to amino acid level,
+#' taking only variable positions under consideration. Than variable amino acid
+#' positions are transformed to counts under \code{inheritance_model} of choice
+#' (this is done with \link{aaVariationToCounts}).
+#'
+#' \code{"expression_levels"} - \code{hla_calls} are transformed to expression
+#' levels using expression dictionaries shipped with package (this is done using
+#' \link{hlaToVariable}). The expression levels from both alleles are than
+#' summed into single variable for each translated HLA gene.
+#'
+#' \code{"allele_g_group"} - \code{hla_calls} are transformed to HLA alleles
+#' groups using G group dictionary shipped with package (this is done using
+#' \link{hlaToVariable}). Than those are transformed to counts under
+#' \code{inheritance_model} of choice (this is done with
+#' \link{hlaCallsToCounts}).
+#'
+#' \code{"allele_supertypes"} - \code{hla_calls} are transformed to HLA alleles
+#' groups using supertypes dictionary shipped with package (this is done using
+#' \link{hlaToVariable}). Than those are transformed to counts under
+#' \code{inheritance_model} of choice (this is done with
+#' \link{hlaCallsToCounts}).
+#'
+#' \code{"allele_groups"} - \code{hla_calls} are transformed to HLA alleles
+#' groups using Bw4/6 and C1/2 groups dictionaries shipped with package (this is
+#' done using \link{hlaToVariable}). Than those are transformed to counts under
+#' \code{inheritance_model} of choice (this is done with
+#' \link{hlaCallsToCounts}).
+#'
+#' \code{"custom"} - will not transform \code{hla_calls} and only joins it with
+#' additional data(\code{...}).
+#'
+#' @return Data frame containing prepared data.
+#'
+#' @examples
+#' hla_calls_file <- system.file("extdata", "HLAHD_output_example.txt", package = "MiDAS")
+#' hla_calls <- readHlaCalls(hla_calls_file)
+#' pheno_file <- system.file("extdata", "pheno_example.txt", package = "MiDAS")
+#' pheno <- read.table(pheno_file, header = TRUE)
+#' covar_file <- system.file("extdata", "covar_example.txt", package = "MiDAS")
+#' covar <- read.table(covar_file, header = TRUE)
+#' prepareMiDASData(hla_calls, pheno, covar, analysis_type = "expression_levels")
+#'
+#' @importFrom assertthat assert_that is.flag is.string see_if
+#' @importFrom dplyr funs group_by left_join mutate summarise_all syms
+#' @importFrom magrittr %>% %<>%
+#' @importFrom rlang .data !!!
+#' @importFrom tidyr gather spread
+#'
+#' @export
+prepareMiDASData <- function(hla_calls,
+                             ...,
+                             analysis_type = c("hla_allele", "aa_level", "expression_levels", "allele_g_group", "allele_supertypes", "allele_groups", "custom"),
+                             inheritance_model = "additive",
+                             indels = TRUE,
+                             unkchar = FALSE
+                             ) {
+  assert_that(
+    checkHlaCallsFormat(hla_calls),
+    is.string(analysis_type),
+    stringMatches(
+      x = analysis_type,
+      choice = c("hla_allele", "aa_level", "expression_levels", "allele_g_group", "allele_supertypes", "allele_groups", "custom")
+    ),
+    is.string(inheritance_model),
+    stringMatches(
+      x = inheritance_model,
+      choice = c("dominant", "recessive", "additive")
+    ),
+    is.flag(indels),
+    is.flag(unkchar)
+  )
+
+  additional_data <- list(...)
+  for (additional_data_frame in additional_data) { # rewrite those tests to addhere to current standard
+    assert_that(
+      checkAdditionalData(additional_data_frame, hla_calls, accept.null = FALSE)
+    )
+  }
+
+  # Process hla_calls based on analysis type
+  if (analysis_type == "hla_allele") {
+    midas_data <- hlaCallsToCounts(
+      hla_calls = hla_calls,
+      inheritance_model = inheritance_model
+    )
+
+  } else if (analysis_type == "aa_level") {
+    midas_data <- hlaToAAVariation(
+      hla_calls = hla_calls,
+      indels = indels,
+      unkchar = unkchar
+    ) %>%
+      aaVariationToCounts(inheritance_model = inheritance_model)
+
+  } else if (analysis_type == "expression_levels") {
+    lib <- listMiDASDictionaries()
+    lib <- grep("expression", lib, value = TRUE)
+    midas_data <- Reduce(
+      f = function(...) left_join(..., by = "ID"),
+      x = lapply(lib, hlaToVariable, hla_calls = hla_calls)
+    )
+
+    assert_that(
+      ncol(midas_data) > 1,
+      msg = "no expression levels were found for input hla_calls"
+    )
+
+    midas_data %<>%
+      gather("expression", "value", -c("ID")) %>%
+      mutate(expression = gsub("_.*", "", .data$expression)) %>%
+      group_by(!!! syms(c("ID", "expression"))) %>%
+      summarise_all(funs(sum)) %>%
+      spread(.data$expression, .data$value, sep = "_")
+
+  } else if (analysis_type == "allele_g_group") {
+    lib <- "4digit_allele_Ggroup"
+    midas_data <- hlaToVariable(hla_calls = hla_calls, dictionary = lib)
+
+    assert_that(
+      ncol(midas_data) > 1,
+      msg = "no allele could be assigned to G group for input hla_calls"
+    )
+
+    midas_data <- hlaCallsToCounts(
+      hla_calls = midas_data,
+      inheritance_model = inheritance_model
+    )
+
+  } else if (analysis_type == "allele_supertypes") {
+    lib <- "4digit_supertype"
+    midas_data <- hlaToVariable(hla_calls = hla_calls, dictionary = lib)
+
+
+    assert_that(
+      ncol(midas_data) > 1,
+      msg = "no allele could be assigned to supertype for input hla_calls"
+    )
+
+    midas_data <- hlaCallsToCounts(
+      hla_calls = midas_data,
+      inheritance_model = inheritance_model,
+      check_hla_format = FALSE
+    ) %>%
+      select(-"Unclassified")
+
+  } else if (analysis_type == "allele_groups") {
+    lib <- c("4digit_B-allele_Bw", "4digit_C-allele_C1-2")
+    midas_data <- Reduce(
+      f = function(...) left_join(..., by = "ID"),
+      x = lapply(lib, hlaToVariable, hla_calls = hla_calls)
+    )
+
+    assert_that(
+      ncol(midas_data) > 1,
+      msg = "no allele could be assigned to allele groups for input hla_calls"
+    )
+
+    midas_data <- hlaCallsToCounts(
+      hla_calls = midas_data,
+      inheritance_model = inheritance_model,
+      check_hla_format = FALSE
+    )
+
+  } else {
+    midas_data <- hla_calls
+  }
+
+  # join with additional_data
+  if (length(additional_data) != 0) {
+    midas_data <- Reduce(
+      f = function(...) left_join(..., by = "ID"),
+      x = additional_data,
+      init = midas_data
+    )
+  }
+
+  return(midas_data)
 }