+++ MiDAS/tests/testthat/test_MiDAS_utils.R
@@ -244,6 +244,17 @@ test_that("string matches", {
 
   expect_error(
     assertthat::assert_that(stringMatches("foo", c("bar", "Foo"))),
-    "\"foo\" should be one of c\\(\"bar\", \"Foo\"\\)."
+    "\"foo\" should be one of \"bar\", \"Foo\"."
+  )
+})
+
+test_that("is flag or null", {
+  expect_equal(isFlagOrNULL(TRUE), TRUE)
+  expect_equal(isFlagOrNULL(NULL), TRUE)
+  # expect_equal(isFlagOrNULL(NA), FALSE) # returns TRUE... strange?
+
+  expect_error(
+    assertthat::assert_that(isFlagOrNULL(1)),
+    "1 is not a flag \\(a length one logical vector\\) or NULL."
   )
 })